Restyle the layout of information section.
