+++ src/components/StrategyPage/Graph/AreaChart.tsx
@@ -23,8 +23,10 @@ const AreaChart = (props: GraphProps) => {
       </GraphTitle>
       <GraphCard>
         <Line
+          height={190}
           data={data}
           options={{
+            maintainAspectRatio: false,
             legend: {
               display: false,
             },
